Function avr_libc::bsearch

pub unsafe extern "C" fn bsearch(
    __key: *const c_void, 
    __base: *const c_void, 
    __nmemb: size_t, 
    __size: size_t, 
    __compar: Option<unsafe extern "C" fn(arg1: *const c_void, arg2: *const c_void) -> i16>
) -> *mut c_void

The bsearch() function searches an array of \c nmemb objects, the initial member of which is pointed to by \c base, for a member that matches the object pointed to by \c key. The size of each member of the array is specified by \c size.

The contents of the array should be in ascending sorted order according to the comparison function referenced by \c compar. The \c compar routine is expected to have two arguments which point to the key object and to an array member, in that order, and should return an integer less than, equal to, or greater than zero if the key object is found, respectively, to be less than, to match, or be greater than the array member.

The bsearch() function returns a pointer to a matching member of the array, or a null pointer if no match is found. If two members compare as equal, which member is matched is unspecified.