Function avr_libc::strtol

pub unsafe extern "C" fn strtol(
    __nptr: *const i8, 
    __endptr: *mut *mut i8, 
    __base: i16
) -> i32

The strtol() function converts the string in \c nptr to a long value. The conversion is done according to the given base, which must be between 2 and 36 inclusive, or be the special value 0.

The string may begin with an arbitrary amount of white space (as determined by isspace()) followed by a single optional \c '+' or \c '-' sign. If \c base is zero or 16, the string may then include a \c "0x" prefix, and the number will be read in base 16; otherwise, a zero base is taken as 10 (decimal) unless the next character is \c '0', in which case it is taken as 8 (octal).

The remainder of the string is converted to a long value in the obvious manner, stopping at the first character which is not a valid digit in the given base. (In bases above 10, the letter \c 'A' in either upper or lower case represents 10, \c 'B' represents 11, and so forth, with \c 'Z' representing 35.)

If \c endptr is not NULL, strtol() stores the address of the first invalid character in \c *endptr. If there were no digits at all, however, strtol() stores the original value of \c nptr in \c endptr. (Thus, if \c *nptr is not \c '\0' but \c **endptr is \c '\0' on return, the entire string was valid.)

The strtol() function returns the result of the conversion, unless the value would underflow or overflow. If no conversion could be performed, 0 is returned. If an overflow or underflow occurs, \c errno is set to \ref avr_errno "ERANGE" and the function return value is clamped to \c LONG_MIN or \c LONG_MAX, respectively.